[修复] 1. 修复PermissionService和TacticsFileService中的类型转换错误 [修复] 2. 将Fail/Success方法重命名为Failed/Succeeded避免与属性冲突 [修复] 3. 修复DbSet名称Users->TacticsUsers [修复] 4. 修复uint到int的类型转换 [修复] 5. 修复bool类型比较(使用true/false而非1/0)

+++ TactiX_Server/Controllers/TacticsHallController.cs
@@ -173,7 +173,7 @@ public class TacticsHallController : ControllerBase
             ModName = file.ModName,
             DownloadCount = (int)file.DownloadCount,
             LikeCount = (int)file.LikeCount,
-            CurrentVersion = version,
+            CurrentVersion = (int)version,
             CreatedAt = file.CreatedAt,
             UpdatedAt = file.UpdatedAt
         };
